# Line wrapping optimization. Fixes #25