--- a/Cargo.toml
+++ b/Cargo.toml
@@ -12,6 +12,7 @@ license = "MIT"
 
 [dependencies]
 byteorder = "1.0.0"
+num = "0.1.37"
 
 [dev-dependencies]
 rand = "=0.3.15"