--- a/src/lib.rs
+++ b/src/lib.rs
@@ -1,10 +1,13 @@
 extern crate byteorder;
 
-use std::{fmt, error, str};
+use std::{fmt, error, ptr, str};
 
 use byteorder::{BigEndian, ByteOrder};
 
 mod tables;
+
+mod line_wrap;
+use line_wrap::{line_wrap_parameters, line_wrap};
 
 /// Available encoding character sets
 #[derive(Clone, Copy, Debug)]
@@ -15,15 +18,41 @@ pub enum CharacterSet {
     UrlSafe
 }
 
+impl CharacterSet {
+    fn encode_table(&self) -> &'static [u8; 64] {
+        match self {
+            &CharacterSet::Standard => tables::STANDARD_ENCODE,
+            &CharacterSet::UrlSafe => tables::URL_SAFE_ENCODE
+        }
+    }
+
+    fn decode_table(&self) -> &'static [u8; 256] {
+        match self {
+            &CharacterSet::Standard => tables::STANDARD_DECODE,
+            &CharacterSet::UrlSafe => tables::URL_SAFE_DECODE
+        }
+    }
+}
+
 #[derive(Clone, Copy, Debug)]
 pub enum LineEnding {
     LF,
     CRLF,
 }
 
+impl LineEnding {
+    fn len(&self) -> usize {
+        match self {
+            &LineEnding::LF => 1,
+            &LineEnding::CRLF => 2
+        }
+    }
+}
+
 #[derive(Clone, Copy, Debug)]
 pub enum LineWrap {
     NoWrap,
+    // wrap length is always > 0
     Wrap(usize, LineEnding)
 }
 
@@ -171,7 +200,7 @@ pub fn decode<T: ?Sized + AsRef<[u8]>>(input: &T) -> Result<Vec<u8>, DecodeError
 ///}
 ///```
 pub fn encode_config<T: ?Sized + AsRef<[u8]>>(input: &T, config: Config) -> String {
-    let mut buf = String::with_capacity(encoded_size(input.as_ref().len(), config));
+    let mut buf = String::with_capacity(encoded_size(input.as_ref().len(), &config));
 
     encode_config_buf(input, config, &mut buf);
 
@@ -179,23 +208,34 @@ pub fn encode_config<T: ?Sized + AsRef<[u8]>>(input: &T, config: Config) -> Stri
 }
 
 /// calculate the base64 encoded string size, including padding
-fn encoded_size(bytes_len: usize, config: Config) -> usize {
+fn encoded_size(bytes_len: usize, config: &Config) -> usize {
+    let msg = "Encoded size exceeds usize";
     let rem = bytes_len % 3;
 
     let complete_input_chunks = bytes_len / 3;
-    let complete_output_chars = complete_input_chunks * 4;
-    let printing_output_chars = if rem == 0 {
-        complete_output_chars
+    let complete_chunk_output = complete_input_chunks.checked_mul(4).expect(msg);
+
+    let encoded_len_no_wrap = if rem > 0 {
+        if config.pad {
+            complete_chunk_output.checked_add(4).expect(msg)
+        } else {
+            let encoded_rem = match rem {
+                1 => 2,
+                2 => 3,
+                _ => panic!("Impossible remainder")
+            };
+            complete_chunk_output.checked_add(encoded_rem).expect(msg)
+        }
     } else {
-        complete_output_chars + 4
-    };
-    let line_ending_output_chars = match config.line_wrap {
-        LineWrap::NoWrap => 0,
-        LineWrap::Wrap(n, LineEnding::CRLF) => printing_output_chars / n * 2,
-        LineWrap::Wrap(n, LineEnding::LF) => printing_output_chars / n,
+        complete_chunk_output
     };
 
-    return printing_output_chars + line_ending_output_chars;
+    match config.line_wrap {
+        LineWrap::NoWrap => encoded_len_no_wrap,
+        LineWrap::Wrap(line_len, line_ending) => {
+            line_wrap_parameters(encoded_len_no_wrap, line_len, line_ending).total_len
+        }
+    }
 }
 
 ///Encode arbitrary octets as base64.
@@ -218,105 +258,140 @@ fn encoded_size(bytes_len: usize, config: Config) -> usize {
 ///```
 pub fn encode_config_buf<T: ?Sized + AsRef<[u8]>>(input: &T, config: Config, buf: &mut String) {
     let input_bytes = input.as_ref();
-    let ref charset = match config.char_set {
-        CharacterSet::Standard => tables::STANDARD_ENCODE,
-        CharacterSet::UrlSafe => tables::URL_SAFE_ENCODE,
-    };
 
     // reserve to make sure the memory we'll be writing to with unsafe is allocated
-    buf.reserve(encoded_size(input_bytes.len(), config));
+    let encoded_size = encoded_size(input_bytes.len(), &config);
+    buf.reserve(encoded_size);
 
     let orig_buf_len = buf.len();
-    let mut fast_loop_output_buf_len = orig_buf_len;
-
-    let input_chunk_len = 6;
-
-    let last_fast_index = input_bytes.len().saturating_sub(8);
 
     // we're only going to insert valid utf8
-    let mut raw = unsafe { buf.as_mut_vec() };
-    // start at the first free part of the output buf
-    let mut output_ptr = unsafe { raw.as_mut_ptr().offset(orig_buf_len as isize) };
+    let mut buf_bytes;
+    unsafe {
+        buf_bytes = buf.as_mut_vec();
+        // expand the string's vec to use its reserved size
+        buf_bytes.set_len(orig_buf_len + encoded_size);
+    }
+
+    let output_bytes_written = {
+        let mut b64_output = &mut buf_bytes[orig_buf_len..];
+
+        // write into the newly reserved space
+        let b64_bytes_written = encode_to_slice(input_bytes, b64_output,
+                                                    config.char_set.encode_table());
+
+        let padding_bytes = if config.pad {
+            add_padding(input_bytes.len(), &mut b64_output[b64_bytes_written..])
+        } else {
+            0
+        };
+
+        let wrappable_bytes = b64_bytes_written + padding_bytes;
+
+        let line_ending_bytes = match config.line_wrap {
+            LineWrap::Wrap(line_len, line_end) =>
+                line_wrap(b64_output, wrappable_bytes, line_len, line_end),
+            LineWrap::NoWrap => 0
+        };
+
+        wrappable_bytes + line_ending_bytes
+    };
+
+    unsafe {
+        buf_bytes.set_len(orig_buf_len + output_bytes_written);
+    }
+}
+
+/// Encode input bytes to utf8 base64 bytes. Does not pad or line wrap.
+/// `output` must be long enough to hold the encoded `input` without padding or line wrapping.
+/// Returns the number of bytes written.
+#[inline]
+fn encode_to_slice(input: &[u8], output: &mut [u8], encode_table: &[u8; 64]) -> usize {
     let mut input_index: usize = 0;
-    if input_bytes.len() >= 8 {
+    let mut output_ptr = output.as_mut_ptr();
+
+    let last_fast_index = input.len().saturating_sub(8);
+    let fast_chunk_len = 6;
+
+    if last_fast_index > 0 {
         while input_index <= last_fast_index {
-            let input_chunk = BigEndian::read_u64(&input_bytes[input_index..(input_index + 8)]);
+            let input_chunk = BigEndian::read_u64(&input[input_index..(input_index + 8)]);
 
             // strip off 6 bits at a time for the first 6 bytes
             unsafe {
-                std::ptr::write(output_ptr, charset[((input_chunk >> 58) & 0x3F) as usize]);
-                std::ptr::write(output_ptr.offset(1), charset[((input_chunk >> 52) & 0x3F) as usize]);
-                std::ptr::write(output_ptr.offset(2), charset[((input_chunk >> 46) & 0x3F) as usize]);
-                std::ptr::write(output_ptr.offset(3), charset[((input_chunk >> 40) & 0x3F) as usize]);
-                std::ptr::write(output_ptr.offset(4), charset[((input_chunk >> 34) & 0x3F) as usize]);
-                std::ptr::write(output_ptr.offset(5), charset[((input_chunk >> 28) & 0x3F) as usize]);
-                std::ptr::write(output_ptr.offset(6), charset[((input_chunk >> 22) & 0x3F) as usize]);
-                std::ptr::write(output_ptr.offset(7), charset[((input_chunk >> 16) & 0x3F) as usize]);
+                ptr::write(output_ptr, encode_table[((input_chunk >> 58) & 0x3F) as usize]);
+                ptr::write(output_ptr.offset(1), encode_table[((input_chunk >> 52) & 0x3F) as usize]);
+                ptr::write(output_ptr.offset(2), encode_table[((input_chunk >> 46) & 0x3F) as usize]);
+                ptr::write(output_ptr.offset(3), encode_table[((input_chunk >> 40) & 0x3F) as usize]);
+                ptr::write(output_ptr.offset(4), encode_table[((input_chunk >> 34) & 0x3F) as usize]);
+                ptr::write(output_ptr.offset(5), encode_table[((input_chunk >> 28) & 0x3F) as usize]);
+                ptr::write(output_ptr.offset(6), encode_table[((input_chunk >> 22) & 0x3F) as usize]);
+                ptr::write(output_ptr.offset(7), encode_table[((input_chunk >> 16) & 0x3F) as usize]);
                 output_ptr = output_ptr.offset(8);
             }
 
-            input_index += input_chunk_len;
-            fast_loop_output_buf_len += 8;
+            input_index += fast_chunk_len;
         }
     }
 
-    unsafe {
-        // expand len to include the bytes we just wrote
-        raw.set_len(fast_loop_output_buf_len);
-    }
+    // Encode the 0 to 7 bytes left after the fast loop.
 
-    // encode the 0 to 7 bytes left after the fast loop
-
-    let rem = input_bytes.len() % 3;
-    let start_of_rem = input_bytes.len() - rem;
+    let rem = input.len() % 3;
+    let start_of_rem = input.len() - rem;
+    let slow_chunk_len = 3;
 
     // start at the first index not handled by fast loop, which may be 0.
-    let mut leftover_index = input_index;
 
-    while leftover_index < start_of_rem {
-        raw.push(charset[(input_bytes[leftover_index] >> 2) as usize]);
-        raw.push(charset[((input_bytes[leftover_index] << 4 | input_bytes[leftover_index + 1] >> 4) & 0x3f) as usize]);
-        raw.push(charset[((input_bytes[leftover_index + 1] << 2 | input_bytes[leftover_index + 2] >> 6) & 0x3f) as usize]);
-        raw.push(charset[(input_bytes[leftover_index + 2] & 0x3f) as usize]);
-
-        leftover_index += 3;
+    while input_index < start_of_rem {
+        unsafe {
+            ptr::write(output_ptr,
+                       encode_table[(input[input_index] >> 2) as usize]);
+            ptr::write(output_ptr.offset(1),
+                       encode_table[((input[input_index] << 4 | input[input_index + 1] >> 4) & 0x3f) as usize]);
+            ptr::write(output_ptr.offset(2),
+                       encode_table[((input[input_index + 1] << 2 | input[input_index + 2] >> 6) & 0x3f) as usize]);
+            ptr::write(output_ptr.offset(3),
+                       encode_table[(input[input_index + 2] & 0x3f) as usize]);
+            output_ptr = output_ptr.offset(4);
+        }
+        input_index += slow_chunk_len;
     }
 
     if rem == 2 {
-        raw.push(charset[(input_bytes[start_of_rem] >> 2) as usize]);
-        raw.push(charset[((input_bytes[start_of_rem] << 4 | input_bytes[start_of_rem + 1] >> 4) & 0x3f) as usize]);
-        raw.push(charset[(input_bytes[start_of_rem + 1] << 2 & 0x3f) as usize]);
+        unsafe {
+            ptr::write(output_ptr,
+                       encode_table[(input[start_of_rem] >> 2) as usize]);
+            ptr::write(output_ptr.offset(1),
+                       encode_table[((input[start_of_rem] << 4 | input[start_of_rem + 1] >> 4) & 0x3f) as usize]);
+            ptr::write(output_ptr.offset(2),
+                       encode_table[(input[start_of_rem + 1] << 2 & 0x3f) as usize]);
+            output_ptr = output_ptr.offset(3);
+        }
     } else if rem == 1 {
-        raw.push(charset[(input_bytes[start_of_rem] >> 2) as usize]);
-        raw.push(charset[(input_bytes[start_of_rem] << 4 & 0x3f) as usize]);
-    }
-
-    if config.pad {
-        for _ in 0..((3 - rem) % 3) {
-            raw.push(0x3d);
+        unsafe {
+            ptr::write(output_ptr,
+                       encode_table[(input[start_of_rem] >> 2) as usize]);
+            ptr::write(output_ptr.offset(1),
+                       encode_table[(input[start_of_rem] << 4 & 0x3f) as usize]);
+            output_ptr = output_ptr.offset(2);
         }
     }
 
-    //TODO FIXME this does the wrong thing for nonempty buffers
-    if orig_buf_len == 0 {
-        if let LineWrap::Wrap(line_size, line_end) = config.line_wrap {
-            let len = raw.len();
-            let mut i = 0;
-            let mut j = 0;
+    // will not underflow: output_ptr is always at least as big as the starting output ptr
+    (output_ptr as usize) - (output.as_ptr() as usize)
+}
 
-            while i < len {
-                if i > 0 && i % line_size == 0 {
-                    match line_end {
-                        LineEnding::LF => { raw.insert(j, b'\n'); j += 1; }
-                        LineEnding::CRLF => { raw.insert(j, b'\r'); raw.insert(j + 1, b'\n'); j += 2; }
-                    }
-                }
-
-                i += 1;
-                j += 1;
-            }
-        }
+/// Write padding characters.
+/// `output` is the slice where padding should be written, of length at least 2.
+fn add_padding(input_len: usize, output: &mut[u8]) -> usize {
+    let rem = input_len % 3;
+    let mut bytes_written = 0;
+    for _ in 0..((3 - rem) % 3) {
+        output[bytes_written] = 0x3d;
+        bytes_written += 1;
     }
+
+    bytes_written
 }
 
 ///Decode from string reference as octets.
@@ -375,10 +450,7 @@ pub fn decode_config_buf<T: ?Sized + AsRef<[u8]>>(input: &T,
         input.as_ref()
     };
 
-    let ref decode_table = match config.char_set {
-        CharacterSet::Standard => tables::STANDARD_DECODE,
-        CharacterSet::UrlSafe => tables::URL_SAFE_DECODE,
-    };
+    let ref decode_table = config.char_set.decode_table();
 
     buffer.reserve(input_bytes.len() * 3 / 4);
 
@@ -574,88 +646,4 @@ pub fn decode_config_buf<T: ?Sized + AsRef<[u8]>>(input: &T,
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn encoded_size_correct() {
-        assert_eq!(0, encoded_size(0, STANDARD));
-
-        assert_eq!(4, encoded_size(1, STANDARD));
-        assert_eq!(4, encoded_size(2, STANDARD));
-        assert_eq!(4, encoded_size(3, STANDARD));
-
-        assert_eq!(8, encoded_size(4, STANDARD));
-        assert_eq!(8, encoded_size(5, STANDARD));
-        assert_eq!(8, encoded_size(6, STANDARD));
-
-        assert_eq!(12, encoded_size(7, STANDARD));
-        assert_eq!(12, encoded_size(8, STANDARD));
-        assert_eq!(12, encoded_size(9, STANDARD));
-
-        assert_eq!(72, encoded_size(54, STANDARD));
-
-        assert_eq!(76, encoded_size(55, STANDARD));
-        assert_eq!(76, encoded_size(56, STANDARD));
-        assert_eq!(76, encoded_size(57, STANDARD));
-
-        assert_eq!(80, encoded_size(58, STANDARD));
-    }
-
-    #[test]
-    fn encoded_size_correct_mime() {
-        assert_eq!(0, encoded_size(0, MIME));
-
-        assert_eq!(4, encoded_size(1, MIME));
-        assert_eq!(4, encoded_size(2, MIME));
-        assert_eq!(4, encoded_size(3, MIME));
-
-        assert_eq!(8, encoded_size(4, MIME));
-        assert_eq!(8, encoded_size(5, MIME));
-        assert_eq!(8, encoded_size(6, MIME));
-
-        assert_eq!(12, encoded_size(7, MIME));
-        assert_eq!(12, encoded_size(8, MIME));
-        assert_eq!(12, encoded_size(9, MIME));
-
-        assert_eq!(72, encoded_size(54, MIME));
-
-        assert_eq!(78, encoded_size(55, MIME));
-        assert_eq!(78, encoded_size(56, MIME));
-        assert_eq!(78, encoded_size(57, MIME));
-
-        assert_eq!(82, encoded_size(58, MIME));
-    }
-
-    #[test]
-    fn encoded_size_correct_lf() {
-        let config = Config::new(
-            CharacterSet::Standard,
-            true,
-            false,
-            LineWrap::Wrap(76, LineEnding::LF)
-        );
-
-        assert_eq!(0, encoded_size(0, config));
-
-        assert_eq!(4, encoded_size(1, config));
-        assert_eq!(4, encoded_size(2, config));
-        assert_eq!(4, encoded_size(3, config));
-
-        assert_eq!(8, encoded_size(4, config));
-        assert_eq!(8, encoded_size(5, config));
-        assert_eq!(8, encoded_size(6, config));
-
-        assert_eq!(12, encoded_size(7, config));
-        assert_eq!(12, encoded_size(8, config));
-        assert_eq!(12, encoded_size(9, config));
-
-        assert_eq!(72, encoded_size(54, config));
-
-        assert_eq!(77, encoded_size(55, config));
-        assert_eq!(77, encoded_size(56, config));
-        assert_eq!(77, encoded_size(57, config));
-
-        assert_eq!(81, encoded_size(58, config));
-    }
-}
+mod tests;