Reimplements line wrapping to use a linear algorithm instead of a quadratic one. By 3 KiB, it was 17x faster (see benchmarks below). I didn't check in a 1MiB benchmark because it took *forever* to run on the old code, but it's 690e3 ns after optimization and 1e9 before, about 1400x faster. 

Initial commit (only added benchmarks, namely the `_mime` ones):

```
test encode_100b                ... bench:          89 ns/iter (+/- 0) = 1123 MB/s
test encode_100b_reuse_buf      ... bench:          71 ns/iter (+/- 0) = 1408 MB/s
test encode_10mib               ... bench:   9,008,847 ns/iter (+/- 308,063) = 1163 MB/s
test encode_10mib_reuse_buf     ... bench:   5,681,881 ns/iter (+/- 9,253) = 1845 MB/s
test encode_30mib               ... bench:  27,901,401 ns/iter (+/- 165,404) = 1127 MB/s
test encode_30mib_reuse_buf     ... bench:  17,047,398 ns/iter (+/- 1,735,299) = 1845 MB/s
test encode_3b                  ... bench:          32 ns/iter (+/- 0) = 93 MB/s
test encode_3b_reuse_buf        ... bench:          13 ns/iter (+/- 0) = 230 MB/s
test encode_3kib                ... bench:       1,707 ns/iter (+/- 3) = 1799 MB/s
test encode_3kib_reuse_buf      ... bench:       1,689 ns/iter (+/- 4) = 1818 MB/s
test encode_3kib_reuse_buf_mime ... bench:      33,256 ns/iter (+/- 66) = 92 MB/s
test encode_3mib                ... bench:   2,454,545 ns/iter (+/- 26,395) = 1281 MB/s
test encode_3mib_reuse_buf      ... bench:   1,693,286 ns/iter (+/- 4,663) = 1857 MB/s
test encode_500b                ... bench:         309 ns/iter (+/- 2) = 1618 MB/s
test encode_500b_reuse_buf      ... bench:         291 ns/iter (+/- 1) = 1718 MB/s
test encode_500b_reuse_buf_mime ... bench:       5,115 ns/iter (+/- 29) = 97 MB/s
test encode_50b                 ... bench:          59 ns/iter (+/- 0) = 847 MB/s
test encode_50b_reuse_buf       ... bench:          41 ns/iter (+/- 0) = 1219 MB/s
```

After optimization:

```
test encode_100b                ... bench:          84 ns/iter (+/- 0) = 1190 MB/s
test encode_100b_reuse_buf      ... bench:          70 ns/iter (+/- 0) = 1428 MB/s
test encode_10mib               ... bench:   9,088,328 ns/iter (+/- 132,041) = 1153 MB/s
test encode_10mib_reuse_buf     ... bench:   5,686,001 ns/iter (+/- 41,927) = 1844 MB/s
test encode_30mib               ... bench:  28,056,341 ns/iter (+/- 158,054) = 1121 MB/s
test encode_30mib_reuse_buf     ... bench:  17,019,695 ns/iter (+/- 20,809) = 1848 MB/s
test encode_3b                  ... bench:          27 ns/iter (+/- 0) = 111 MB/s
test encode_3b_reuse_buf        ... bench:          12 ns/iter (+/- 0) = 250 MB/s
test encode_3kib                ... bench:       1,699 ns/iter (+/- 6) = 1808 MB/s
test encode_3kib_reuse_buf      ... bench:       1,683 ns/iter (+/- 5) = 1825 MB/s
test encode_3kib_reuse_buf_mime ... bench:       1,951 ns/iter (+/- 16) = 1574 MB/s
test encode_3mib                ... bench:   2,469,405 ns/iter (+/- 34,354) = 1273 MB/s
test encode_3mib_reuse_buf      ... bench:   1,690,765 ns/iter (+/- 4,259) = 1860 MB/s
test encode_500b                ... bench:         305 ns/iter (+/- 1) = 1639 MB/s
test encode_500b_reuse_buf      ... bench:         290 ns/iter (+/- 0) = 1724 MB/s
test encode_500b_reuse_buf_mime ... bench:         355 ns/iter (+/- 1) = 1408 MB/s
test encode_50b                 ... bench:          55 ns/iter (+/- 0) = 909 MB/s
test encode_50b_reuse_buf       ... bench:          40 ns/iter (+/- 0) = 1250 MB/s
```