@marshallpierce 2017-05-03
No problem, I merged in your changes and added a few other `checked_add()`'s. Most of the line wrapping logic was already pretty laden with `checked_*` fortunately.