@marshallpierce 2017-05-06
The `minimize-unsafe` branch (see #29) builds on top of this to remove usage of unsafe. If you'd rather I can just PR that branch and close this one.