@alicemaz 2017-05-14
I think we can merge in `minimize-unsafe` and proceed from there