@marshallpierce 2017-05-14
Sounds good; closing in favor of #31.